Edited Main Menu, Swapped to using toon car, added credit scene
I duplicated the Main Menu to add a background type feel to it so it isn't just blue behind it. I just used the green toon car for the player and edited the material to use blue, red, and yellow for the AI. I also added a Credits scene that will need to be added to probably. I am pretty sure we need a controls page as well.