Updated End Screen and some of the code. To make it a little simpler I think.
I made the minutes and seconds get stored in the static GameManager script at the end of the race so instead of loading the endscreen from the RollMovement script the RollMovement scrip now just sets a static bool of raceFinished to true. Then the GUIManager script stores the minutes and seconds and then loads the EndScreen from that class. Will work on getting the sound working now.