Added in the sounds and made a few adjustments
I added in the sounds. Made the cars have Idle engine sounds running engine sounds as well as added the crash sound to each car so when a car collides with anything at a certain speed it plays that. Added the background music to a object and used Micheal's singleton script to keep one and only one from the main menu. And added the gunshot to the GUIManager to play when starting race.

I adjusted the terrain and world decompose because I noticed that if you got enough left tracks or right tracks you could make it outside the decomposed area.

I also changed the stickiness of the players tires to allow them to spin out slightly. This was to make the AI a little more competitive. But they are still easily beaten.